requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
python-docx>=1.1.0
lxml>=5.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.0
browser-cookie3>=0.19.1
selenium>=4.15.0
